Rename submod
